+++ src/main/java/io/github/pitzzahh/util/utilities/classes/enums/Difficulty.java
@@ -1,0 +1,19 @@
+package io.github.pitzzahh.util.utilities.classes.enums;
+
+/**
+ * Enum for difficult situations.
+ */
+public enum Difficulty {
+    /**
+     * Situation is easy.
+     */
+    EASY,
+    /**
+     * Situation is normal.
+     */
+    MEDIUM,
+    /**
+     * Situation is hard to bear.
+     */
+    HARD
+}
