+++ src/main/java/io/github/pitzzahh/util/utilities/classes/enums/Role.java
@@ -1,0 +1,51 @@
+package io.github.pitzzahh.util.utilities.classes.enums;
+
+/**
+ * enum that can be used for making roles.
+ */
+public enum Role {
+    /**
+     * Standard role.
+     */
+    STANDARD,
+    /**
+     * Role for a user.
+     */
+    USER,
+    /**
+     * Role for a admin
+     */
+    ADMIN,
+    /**
+     * Role for a client.
+     */
+    CLIENT,
+    /**
+     * Role for a guests.
+     */
+    GUEST,
+    /**
+     * Role for a student.
+     */
+    STUDENT,
+    /**
+     * Role for a teacher
+     */
+    TEACHER,
+    /**
+     * Role for a staff.
+     */
+    STAFF,
+    /**
+     * Role for a player.
+     */
+    PLAYER,
+    /**
+     * Role for a computer.
+     */
+    COMPUTER,
+    /**
+     * Role for a human.
+     */
+    HUMAN
+}
