--- src/main/java/io/github/pitzzahh/utilities/classes/enums/Quadrant.java
@@ -1,32 +1,0 @@
-package io.github.pitzzahh.utilities.classes.enums;
-
-/**
- * Used in representing which quadrant the coordinate belongs.
- */
-public enum Quadrant {
-
-    /**
-     * If lies in the first quadrant
-     */
-    FIRST_QUADRANT,
-
-    /**
-     * If lies in the second quadrant
-     */
-    SECOND_QUADRANT,
-
-    /**
-     * If lies in the third quadrant
-     */
-    THIRD_QUADRANT,
-
-    /**
-     * If lies in the fourth quadrant
-     */
-    FOURTH_QUADRANT,
-
-    /**
-     * If lies at the origin.
-     */
-    ORIGIN
-}
