--- src/main/java/io/github/pitzzahh/utilities/classes/enums/Status.java
@@ -1,19 +1,0 @@
-package io.github.pitzzahh.utilities.classes.enums;
-
-/**
- * enum used in processes.
- */
-public enum Status {
-    /**
-     * Denotes that a process turned out successful
-     */
-    SUCCESS,
-    /**
-     * Denotes that a process turned out error
-     */
-    ERROR,
-    /**
-     * Denotes that a process cannot be performed.
-     */
-    CANNOT_PERFORM_OPERATION
-}
